Add pollfd_iterator_invalided, my_close, remove std::map.find
pollfd_iterator is thrown when Server.fds is invalided and caught in the loop to restart it.

When revent is handled, it is set to zero to avoid handling it again.

my_close throws std::runtime_error on -1.

std::map.find is not used since iterator invalidation is now handled properly